aula_3/taskbot_ai/pyproject.toml
[project]
name = "taskbot-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian Vieira Salgado",email = "vieira4ev@gmail.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "langgraph (>=0.5.2,<0.6.0)",
    "langgraph-prebuilt (>=0.5.2,<0.6.0)",
    "langchain-core (>=0.3.68,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-telegram-bot (>=22.2,<23.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.176.0,<3.0.0)",
    "google-api-core (>=2.25.1,<3.0.0)"
]

[tool.poetry]
packages = [
    {include = "agent", from = "src"},
    {include = "bot", from = "src"},
    {include = "tasks", from = "src"},
    {include = "tools", from = "src"}
]


[tool.poetry.group.dev.dependencies]
langgraph-cli = {extras = ["inmem"], version = "^0.3.4"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
